refactor(analytics): migrate price fields from Double to BigDecimal, wire monthly stats to live API
- Change price/revenue fields from Double to BigDecimal (backend +
  entity + DTOs) to fix floating point precision artifacts
  (e.g. 569.0899999999999)
- Wire Analytics component to GET /api/business/{businessId} for
  totals and monthly stats, replacing hardcoded revenueData
- Apply fillMonths() to backfill sparse monthly API data to full
  Jan-Dec range for consistent chart rendering
- Update "6-month trend" label mismatch (still pending, see TODO)

TODO: serviceData (pie chart) still hardcoded mock, needs real
  service-distribution endpoint
TODO: hourlyData (peak hours bar chart) still hardcoded mock, needs
  real hourly-booking-distribution endpoint
TODO: avg. rating stat still hardcoded "4.8", needs wiring to
  ServiceReview aggregation query

+++ demo/src/main/java/com/example/demo/dto/response/BusinessTotalsDto.java
@@ -1,7 +1,9 @@
 package com.example.demo.dto.response;
 
+import java.math.BigDecimal;
+
 public record BusinessTotalsDto(
-    double totalRevenue,
+    BigDecimal totalRevenue,
     Long totalBookings,
     Long totalCustomer
 ) {}
